requirements.txt
Django==1.7.10
dj-database-url==0.2.1
psycopg2==2.5.1
wsgiref==0.1.2
markdown==2.6
bleach==1.4.1
mdx_del_ins==1.0
mdx_linkify==0.6
pytz==2014.10
djangorestframework==3.0.4
Wand==0.4.1
Pillow==2.7.0
django-imagekit==3.2.5
python-social-auth
future
